Adding matches and games
