+++ src/tournament/match/MatchOptions.ts
@@ -1,0 +1,4 @@
+export default class MatchOptions {
+    maxGames: number;
+    timeout: number;
+}
