+++ src/tournament/match/game/GameEvents.ts
@@ -1,0 +1,3 @@
+export default interface GameEvents {
+    onGameStart(): void
+}
