+++ src/tournament/match/game/GameOptions.ts
@@ -1,0 +1,3 @@
+export default class GameOptions {
+    timeout: number;
+}
