+++ src/tournament/matchmaker/Matchmaker.ts
@@ -1,0 +1,7 @@
+import { TournamentStats } from "../model/TournamentStats";
+import Match from "../match/Match";
+
+export default interface Matchmaker {
+    isFinished(): Boolean
+    getRemainingMatches(stats: TournamentStats): Match[]
+}
