+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "lib": ["es6"],
+    "lib": ["es7"],
     "module": "commonjs",
     "noImplicitAny": true,
     "noEmitOnError": true,
